// src/api/generated/model/settingsNotificationsDtoNotificationStyle.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Workforms API
 * HTTP API (WebSocket chat is not described here)
 * OpenAPI spec version: 1.0
 */

export type SettingsNotificationsDtoNotificationStyle =
  (typeof SettingsNotificationsDtoNotificationStyle)[keyof typeof SettingsNotificationsDtoNotificationStyle];

export const SettingsNotificationsDtoNotificationStyle = {
  push: 'push',
  email: 'email',
  both: 'both',
} as const;